split out widget-related API functions from main api.js file
This just splits the widget-related API functions out into a separate
file, like we've already done for a number of other files.

part of #1667
pr #2020
